requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai >= 1.0.0
pydantic
python-dotenv
typing-extensions
ipykernel
packaging
tavily-python
langchain-tavily
sqlalchemy
google-generativeai
requests
httpx
tiktoken
